Implemented subset-based modifier overrides (don't have good name for this and too lazy to explain but basically, it allows you to combine modifications in a really smart way!)
Also fixed an issue where sometimes a ScrollModifications would still be active after releasing a modifier key that it is preconditioned on

+++ Helper/InputTransformation/Touch/GestureScrollSimulator.h
@@ -14,7 +14,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface GestureScrollSimulator : NSObject
 
-+ (void)postGestureScrollEventWithDeltaX:(double)dx deltaY:(double)dy phase:(IOHIDEventPhaseBits)phase;
++ (void)postGestureScrollEventWithDeltaX:(int64_t)dx deltaY:(int64_t)dy phase:(IOHIDEventPhaseBits)phase;
 + (void)stopMomentumScroll;
 
 @end
